fix: add security to amenity options

+++ src/modules/blogPost/blogPost.resolver.ts
@@ -1,11 +1,11 @@
 import { Resolver, Query, Args, Int, Mutation, Context } from '@nestjs/graphql';
 import { BlogPostService } from './blogPost.service';
-import { BlogPost, BlogPostList } from '@src/models/global.model';
-import { UseGuards } from '@nestjs/common';
+import { BlogPost, BlogPostList, CtxType } from '@src/models/global.model';
+import { ForbiddenException, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@src/guards/auth.guard';
 import { CreateBlogPostInput } from './dtos/create-blog.input';
 import { UpdateBlogPostInput } from './dtos/update-blog.input';
-import { BlogStatus, BlogTags } from '@prisma/client';
+import { BlogStatus, BlogTags, UserType } from '@prisma/client';
 // import { Controller } from '@nestjs/common';
 
 @Resolver(() => BlogPost)
@@ -52,20 +52,50 @@ export class BlogPostResolver {
   @Mutation(() => BlogPost)
   @UseGuards(AuthGuard)
   async createBlogPost(
-    @Context() ctx: any,
+    @Context() ctx: CtxType,
     @Args('data') data: CreateBlogPostInput,
   ) {
     // const userId = Number(ctx.user.sub);
+    // prevent update by other users then writer or superadmin
+
+    if (
+      ctx.user.userType !== UserType.WRITER &&
+      ctx.user.userType !== UserType.SUPERADMIN
+    ) {
+      throw new ForbiddenException('You are not allowed to create blog post');
+    }
+
     return this.blogPostService.createBlogPost(data);
   }
 
   @Mutation(() => BlogPost)
-  async updateBlogPost(@Args('data') data: UpdateBlogPostInput) {
+  @UseGuards(AuthGuard)
+  async updateBlogPost(
+    @Args('data') data: UpdateBlogPostInput,
+    @Context() ctx: CtxType,
+  ) {
+    // prevent update by other users then writer or superadmin
+    if (
+      ctx.user.userType !== UserType.WRITER &&
+      ctx.user.userType !== UserType.SUPERADMIN
+    ) {
+      throw new ForbiddenException('You are not allowed to update blog post');
+    }
+
     return this.blogPostService.updateBlogPost(data.id, data);
   }
 
   @Mutation(() => BlogPost)
-  async deleteBlogPost(@Args('id') id: number) {
+  @UseGuards(AuthGuard)
+  async deleteBlogPost(@Args('id') id: number, @Context() ctx: CtxType) {
+    // prevent delete by other users then writer or superadmin
+    if (
+      ctx.user.userType !== UserType.WRITER &&
+      ctx.user.userType !== UserType.SUPERADMIN
+    ) {
+      throw new ForbiddenException('You are not allowed to delete blog post');
+    }
+
     return this.blogPostService.deleteBlogPost(id);
   }
 }
